HOTFIX: Maltrail Malware DB + Bug Fix #1606 (#1634)
Added support for Maltrail Malware DB
Fixes a bug in libsast #1606
