+++ MobSF/settings.py
@@ -21,7 +21,7 @@ logger = logging.getLogger(__name__)
 # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 
-MOBSF_VER = 'v3.2.3 Beta'
+MOBSF_VER = 'v3.2.4 Beta'
 
 BANNER = """
   __  __       _    ____  _____   _____  ____  
@@ -328,6 +328,8 @@ else:
 
     # -----External URLS--------------------------
     MALWARE_DB_URL = 'https://www.malwaredomainlist.com/mdlcsv.php'
+    MALTRAIL_DB_URL = ('https://raw.githubusercontent.com/stamparm/aux/'
+                       'master/maltrail-malware-domains.txt')
     VIRUS_TOTAL_BASE_URL = 'https://www.virustotal.com/vtapi/v2/file/'
     EXODUS_URL = 'https://reports.exodus-privacy.eu.org'
     APPMONSTA_URL = 'https://api.appmonsta.com/v1/stores/android/details/'
