+++ requirements.txt
@@ -21,7 +21,7 @@ distro==1.5.0
 IP2Location==8.5.1
 lief==0.10.1
 http-tools==1.1.1
-libsast==1.3.5
+libsast==1.3.6
 # Do not update the following
 apkid==2.1.1
 cryptography<3.3,>=3.2 # pyup: ignore
